refactor(ws-client): organize src into transport/ and lib/ folders
