+++ packages/ws-client/src/browser.ts
@@ -1,4 +1,3 @@
-import { createBrowserTransport } from './browser-transport.js'
 import {
   type Awaitable,
   type BrowserWebSocketClientOptions,
@@ -10,6 +9,7 @@ import {
   type DataMode,
   WebSocketConnectionEngine,
 } from './connection.js'
+import { createBrowserTransport } from './transport/browser-transport.js'
 
 /**
  * A single WebSocket connection, consumed as an `AsyncIterable` of messages.
@@ -71,7 +71,7 @@ export {
   SocketError,
   WebSocketClientError,
   WebSocketConnectionError,
-} from './errors.js'
+} from './lib/errors.js'
 export type {
   Awaitable,
   BrowserWebSocketClientOptions,
@@ -82,11 +82,11 @@ export {
   FATAL_CLOSE_CODES,
   defaultShouldReconnect,
   isReconnectableClose,
-} from './reconnect-policy.js'
+} from './lib/reconnect-policy.js'
 export type {
   CloseEventDetail,
   WebSocketClientEventMap,
   WebSocketConnectionEventMap,
-} from './typed-event-target.js'
+} from './lib/typed-event-target.js'
 
-export { CloseCode } from './close-codes.js'
+export { CloseCode } from './lib/close-codes.js'
