+++ packages/ws-client/src/lib/close-codes.ts
