+++ packages/ws-client/src/lib/errors.ts
