+++ packages/ws-client/tests/client.test.ts
@@ -6,8 +6,8 @@ import {
   BufferOverflowError,
   WebSocketClientError,
   WebSocketConnectionError,
-} from '../src/errors.js'
-import type { CloseEventDetail } from '../src/typed-event-target.js'
+} from '../src/lib/errors.js'
+import type { CloseEventDetail } from '../src/lib/typed-event-target.js'
 import { MockTransport } from './_util/mock-transport.js'
 
 function makeClient<M extends 'auto' | 'text' | 'binary' = 'auto'>(
