+++ packages/ws-client/tests/connection-idle.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { WebSocketConnectionEngine } from '../src/connection.js'
-import { IdleTimeoutError } from '../src/errors.js'
+import { IdleTimeoutError } from '../src/lib/errors.js'
 import { MockTransport } from './_util/mock-transport.js'
 
 describe('WebSocketConnectionEngine idle timeout', () => {
